♻️refactor: Type Converter 파일 이동 (#45)
