+++ src/main/java/kr/modernworld/modernworldv2/admin/infrastructure/persistence/converter/ItemTypeConverter.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.admin.infrastructure.persistence.entity;
+package kr.modernworld.modernworldv2.admin.infrastructure.persistence.converter;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
